Separated out mini-schedule styles
