+++ autoscheduler/frontend/src/components/SchedulingPage/SchedulePreview/MiniSchedule/MiniSchedule.css
@@ -1,0 +1,42 @@
+.hour-marker {
+    flex-grow: 1;
+    border-bottom: 1px solid darkgray;
+    height: 0;
+    align-self: stretch;
+    position: relative;
+    top: 0;
+}
+
+.calendar-body {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    background-color: white;
+    height: 100%;
+}
+
+.meetings-container {
+    position: absolute;
+    height: 100%;
+    width: 100%;
+    top: 0px;
+    left: 0px;
+    margin-left: 0px;
+    margin-top: 0px;
+    display: flex;
+    flex-direction: row;
+}
+
+.header {
+    border-top-left-radius: 4px;
+    border-top-right-radius: 4px;
+    height: 8px;
+    background-color: #500000;
+}
+
+.container {
+    width: 100%;
+    align-self: stretch;
+    min-height: 120px;
+    margin: 8px 0px 8px 16px;
+}
