+++ autoscheduler/frontend/src/components/SchedulingPage/SchedulePreview/SchedulePreview.css
@@ -1,5 +1,5 @@
 .gpa {
-    float: right;
+    display: inline-block;
 }
 
 #card-header {
